Shuffle tilde parsing tests, add process_card controller
Ignoring the substring test for parsing tilde rules because the
following regex doesn't seem to work (where <name> is the literal name
of the card):
  \b<name>\b
For ascii this works, but I'm having trouble getting the word boundry
match (\b) to work with unicode characters (\xb2).  Instead I'm opting
for over-matching and supporting unicode.  I'm not sure this has any
actual impact, but I'll open an issue and keep an eye on requests that
include tilde rules searches.