Add rulings, returnable_card_fields
rcf constrain which values can be returned, and in the case of fields
with multiple representations (power, toughness) how they are
returned.  Note that the keys of this dictionary do not necessarily
correspond to the values that a consumer would pass in.  For instance,
'rules' would likely be passed as 'o', 'type' as 't'.